Refactor large files for maintainability and increase test coverage
Split test_sweep.py (1,866 lines) into 9 focused test modules with a
backward-compatible re-export shim. Extract 5 shared helper functions
from orchestration.py to eliminate ~200 lines of duplication between
launch_sweep() and launch_all_stages(). Extract 5 shared helpers from
train_base.py to consolidate duplicated setup logic between train() and
train_curriculum() (alg kwargs, vecnorm loading, model creation,
callbacks, post-training save). Add 25 new tests covering the extracted
helpers and previously untested _dedup_trial_rows.
